--- hyeontaek/twoSum/TwoSum.java
@@ -1,24 +1,0 @@
-package hyeontaek.twoSum;
-
-import java.util.Arrays;
-
-public class TwoSum {
-
-  public int[] twoSum(int[] nums, int target) {
-    for (int i = 0; i < nums.length; i++) {
-      for (int j = i + 1; j < nums.length; j++) {
-        if (nums[i] + nums[j] == target) {
-          return new int[]{i, j};
-        }
-      }
-    }
-    return new int[] {};
-  }
-
-  public static void main(String[] args) {
-    TwoSum solution = new TwoSum();
-    int[] nums = {3,3};
-    int target = 6;
-    System.out.println(Arrays.toString(solution.twoSum(nums, target)));
-  }
-}
